+++ picodi/_scopes.py
@@ -20,37 +20,46 @@ class Scope:
     Inherit from :class:`AutoScope` or :class:`ManualScope`.
     """
 
-    def get(self, key: Hashable) -> Any:
+    def get(self, key: Hashable, *, global_key: Hashable) -> Any:
         """
         Get a value by key.
         If value is not exists must raise KeyError.
 
         :param key: key to get value, typically a dependency function.
+        :param global_key: typically a function that requesting dependencies
         :raises KeyError: if value not exists.
         """
         raise NotImplementedError
 
-    def set(self, key: Hashable, value: Any) -> None:
+    def set(self, key: Hashable, value: Any, *, global_key: Hashable) -> None:
         """
         Set a value by key.
 
         :param key: key to set value, typically a dependency function.
         :param value: value to set, typically a dependency instance.
+        :param global_key: typically a function that requesting dependencies
         """
         raise NotImplementedError
 
-    def enter_inject(self) -> None:
+    def enter_inject(self, global_key: Hashable) -> None:  # noqa: U100
         """
         Called when entering an :func:`inject` decorator.
+
+        :param global_key: typically a function that requesting dependencies
         """
         return None
 
-    def exit_inject(self, exc: BaseException | None = None) -> None:  # noqa: U100
+    def exit_inject(
+        self, exc: BaseException | None = None, *, global_key: Hashable  # noqa: U100
+    ) -> None:
         """
         Called before exiting a :func:`inject` decorator.
 
         ``shutdown`` will be called after this, e.g.:
             ``exit_inject`` -> ``shutdown`` -> ``inject`` wrapper returns.
+
+        :param exc: exception that was raised in the context.
+        :param global_key: typically a function that requesting dependencies
         """
         return None
 
@@ -61,37 +70,7 @@ class AutoScope(Scope):
     after exiting the context.
 
     Don't use this class directly.
-    Usually, you don't want to inherit this class, but you can.
     """
-
-    def enter(
-        self,
-        exit_stack: ExitStack,
-        context_manager: AsyncContextManager | ContextManager,
-    ) -> Awaitable:
-        """
-        Hook for entering yielded dependencies context. Will be called automatically
-        by picodi.
-        Usually you don't need to override this method.
-
-        :param exit_stack: :class:`support.ExitStack` instance. Instance of ExitStack
-            will be created automatically in :func:`inject` decorator.
-        :param context_manager: context manager created from yield dependency.
-        """
-        return exit_stack.enter_context(context_manager)
-
-    def shutdown(
-        self, exit_stack: ExitStack, exc: BaseException | None = None
-    ) -> Awaitable:
-        """
-        Hook for closing dependencies. Will be called automatically by picodi.
-        Usually you don't need to override this method.
-
-        :param exit_stack: :class:`support.ExitStack` instance. Instance of ExitStack
-            will be created automatically in :func:`inject` decorator.
-        :param exc: exception that was raised in the context.
-        """
-        return exit_stack.close(exc)
 
 
 class ManualScope(Scope):
@@ -107,22 +86,29 @@ class ManualScope(Scope):
     """
 
     def enter(
-        self, context_manager: AsyncContextManager | ContextManager  # noqa: U100
+        self,
+        context_manager: AsyncContextManager | ContextManager,  # noqa: U100
+        *,
+        global_key: Hashable,  # noqa: U100
     ) -> Awaitable:
         """
         Hook for entering yielded dependencies context. Will be called automatically
         by picodi or when you call :func:`init_dependencies`.
 
         :param context_manager: context manager created from yield dependency.
+        :param global_key: typically a function that requesting dependencies
         """
         return NullAwaitable()
 
-    def shutdown(self, exc: BaseException | None = None) -> Awaitable:  # noqa: U100
+    def shutdown(
+        self, exc: BaseException | None = None, *, global_key: Hashable  # noqa: U100
+    ) -> Awaitable:
         """
         Hook for shutdown dependencies.
         Will be called when you call :func:`shutdown_dependencies`
 
         :param exc: exception that was raised in the context.
+        :param global_key: typically a function that requesting dependencies
         """
         return NullAwaitable()
 
@@ -137,10 +123,12 @@ class NullScope(AutoScope):
     This is the default scope.
     """
 
-    def get(self, key: Hashable) -> Any:
+    def get(self, key: Hashable, *, global_key: Hashable) -> Any:  # noqa: U100
         raise KeyError(key)
 
-    def set(self, key: Hashable, value: Any) -> None:  # noqa: U100
+    def set(
+        self, key: Hashable, value: Any, *, global_key: Hashable  # noqa: U100
+    ) -> None:
         return None
 
 
@@ -154,16 +142,29 @@ class SingletonScope(ManualScope):
         self._exit_stack = ExitStack()
         self._store: dict[Hashable, Any] = {}
 
-    def get(self, key: Hashable) -> Any:
+    def get(self, key: Hashable, *, global_key: Hashable) -> Any:  # noqa: U100
         return self._store[key]
 
-    def set(self, key: Hashable, value: Any) -> None:
+    def set(
+        self,
+        key: Hashable,
+        value: Any,
+        *,
+        global_key: Hashable,  # noqa: U100
+    ) -> None:
         self._store[key] = value
 
-    def enter(self, context_manager: AsyncContextManager | ContextManager) -> Awaitable:
+    def enter(
+        self,
+        context_manager: AsyncContextManager | ContextManager,
+        *,
+        global_key: Hashable,  # noqa: U100
+    ) -> Awaitable:
         return self._exit_stack.enter_context(context_manager)
 
-    def shutdown(self, exc: BaseException | None = None) -> Awaitable:
+    def shutdown(
+        self, exc: BaseException | None = None, *, global_key: Hashable  # noqa: U100
+    ) -> Awaitable:
         self._store.clear()
         return self._exit_stack.close(exc)
 
@@ -180,7 +181,7 @@ class ContextVarScope(ManualScope):
         )
         self._store: dict[Any, ContextVar[Any]] = {}
 
-    def get(self, key: Hashable) -> Any:
+    def get(self, key: Hashable, *, global_key: Hashable) -> Any:  # noqa: U100
         try:
             value = self._store[key].get()
         except LookupError:
@@ -189,18 +190,31 @@ class ContextVarScope(ManualScope):
             raise KeyError(key)
         return value
 
-    def set(self, key: Hashable, value: Any) -> None:
+    def set(
+        self,
+        key: Hashable,
+        value: Any,
+        *,
+        global_key: Hashable,  # noqa: U100
+    ) -> None:
         try:
             var = self._store[key]
         except KeyError:
             var = self._store[key] = ContextVar("picodi_ContextVarScope_var")
         var.set(value)
 
-    def enter(self, context_manager: AsyncContextManager | ContextManager) -> Awaitable:
+    def enter(
+        self,
+        context_manager: AsyncContextManager | ContextManager,
+        *,
+        global_key: Hashable,  # noqa: U100
+    ) -> Awaitable:
         exit_stack = self._get_exit_stack()
         return exit_stack.enter_context(context_manager)
 
-    def shutdown(self, exc: BaseException | None = None) -> Any:
+    def shutdown(
+        self, exc: BaseException | None = None, *, global_key: Hashable  # noqa: U100
+    ) -> Any:
         for var in self._store.values():
             var.set(unset)
         exit_stack = self._get_exit_stack()